+++ app/components/dashbaord/playerProfileForm.tsx
@@ -1,18 +1,33 @@
 import {DefaultFeedback, Player} from "@prisma/client";
+import InputWithLabel from "~/components/common/form/InputWithLabel";
+import messages from "~/components/i18n/messages";
 import ContentContainer from "~/components/common/container/ContentContainer";
 import Subheading from "~/components/common/header/Subheading";
-import messages from "~/components/i18n/messages";
 import DefaultFeedbackComponent from "~/components/player/feedback/DefaultFeedbackComponent";
 import ButtonContainer from "~/components/common/container/ButtonContainer";
 import DefaultButton from "~/components/common/buttons/DefaultButton";
-import EditProfile from "~/components/dashbaord/editProfile";
 
 
-const PlayerProfile = ({player, defaultFeedback}: { player: Player, defaultFeedback: DefaultFeedback }) => {
+const DashboardPlayerProfileForm = ({player, defaultFeedback}: { player: Player, defaultFeedback: DefaultFeedback }) => {
     return (
         <ContentContainer className={"shadow-lg shadow-indigo-500/50"}>
             <Subheading title={messages.dashboard.playerProfile}/>
-            <EditProfile player={player}/>
+            <div>
+                <InputWithLabel
+                    id={'name'}
+                    type={'text'}
+                    name={'dashboard.profile.player.name'}
+                    defaultValue={player.name}
+                    label={messages.playerProfileForm.name}
+                />
+                <InputWithLabel
+                    id={'mail'}
+                    type={'email'}
+                    name={'dashboard.profile.player.email'}
+                    defaultValue={player.email}
+                    label={messages.playerProfileForm.email}
+                />
+            </div>
             <hr className="my-8 h-px bg-gray-400 border-0"/>
             <DefaultFeedbackComponent defaultFeedback={defaultFeedback}
                                       title={messages.dashboard.playerDefaultStatus}
@@ -26,4 +41,4 @@ const PlayerProfile = ({player, defaultFeedback}: { player: Player, defaultFeedb
     )
 }
 
-export default PlayerProfile
+export default DashboardPlayerProfileForm
